Improve integrity decomp

+++ src/integrity.c
@@ -25,7 +25,7 @@ u32 Integrity_MACOwner_IsBad(void) {
 	u32  offset;
 	
 	// Obfuscated handling of function address
-	addr = (u8*)(((u32)&RunEncrypted_MACOwner_IsBad[ENC_VAL_1]) - (ENC_VAL_1 * 2));
+	addr = (u8*)&RunEncrypted_MACOwner_IsBad[ENC_VAL_1] - (ENC_VAL_1 * 2);
 	addr += INTEGRITY_OBFS_OFFSET;
 	offset = ENC_VAL_1 - INTEGRITY_OBFS_OFFSET;
 	
@@ -59,7 +59,7 @@ u32 Integrity_MACOwner_IsGood(void) {
 	u32  offset;
 	
 	// Obfuscated handling of function address
-	addr = (u8*)(((u32)&RunEncrypted_MACOwner_IsGood[ENC_VAL_1]) - (ENC_VAL_1 * 2));
+	addr = (u8*)&RunEncrypted_MACOwner_IsGood[ENC_VAL_1] - (ENC_VAL_1 * 2);
 	addr += INTEGRITY_OBFS_OFFSET;
 	offset = ENC_VAL_1 - INTEGRITY_OBFS_OFFSET;
 	
@@ -93,7 +93,7 @@ u32 Integrity_ROMTest_IsBad(void) {
 	u32  offset;
 	
 	// Obfuscated handling of function address
-	addr = (u8*)(((u32)&RunEncrypted_ROMTest_IsBad[ENC_VAL_1]) - (ENC_VAL_1 * 2));
+	addr = (u8*)&RunEncrypted_ROMTest_IsBad[ENC_VAL_1] - (ENC_VAL_1 * 2);
 	addr += INTEGRITY_OBFS_OFFSET;
 	offset = ENC_VAL_1 - INTEGRITY_OBFS_OFFSET;
 	
@@ -127,7 +127,7 @@ u32 Integrity_ROMTest_IsGood(void) {
 	u32  offset;
 	
 	// Obfuscated handling of function address
-	addr = (u8*)(((u32)&RunEncrypted_ROMTest_IsGood[ENC_VAL_1]) - (ENC_VAL_1 * 2));
+	addr = (u8*)&RunEncrypted_ROMTest_IsGood[ENC_VAL_1] - (ENC_VAL_1 * 2);
 	addr += INTEGRITY_OBFS_OFFSET;
 	offset = ENC_VAL_1 - INTEGRITY_OBFS_OFFSET;
